+++ src/app/auth/page.tsx
@@ -1,5 +1,10 @@
+import { Suspense } from "react";
 import AuthComponent from "./components/AuthComponent";
 
 export default function page() {
-  return <AuthComponent />;
+  return (
+    <Suspense fallback={null}>
+      <AuthComponent />
+    </Suspense>
+  );
 }
