+++ src/app/http/index.ts
@@ -1,21 +1,29 @@
 import i18n from "@/i18n";
-import axios from "axios";
+import axios, { HttpStatusCode } from "axios";
 import { toast } from "sonner";
 
 export const http = axios.create({
   headers: {
-    "Content-Type": 'application/json"',
+    "Content-Type": "application/json",
   },
 });
 
 http.interceptors.response.use(
   (response) => response,
   (error) => {
-    if (error.response?.status === 500) {
-      toast.error(
-        i18n?.t("error_occurred") ||
-          "An unexpected error occurred. Please try again later."
-      );
+    if (error.response) {
+      const { status } = error.response;
+
+      if (status === HttpStatusCode.InternalServerError) {
+        toast.error(
+          i18n?.t("error_occurred") ||
+            "An unexpected error occurred. Please try again later."
+        );
+      }
+
+      if (status === HttpStatusCode.Unauthorized) {
+        window.location.href = `/auth?error=unauthorized`;
+      }
     }
     return Promise.reject(error);
   }
